Split `EdgeCache` and `NodeCache` out into separate files.
